pyproject.toml
[project]
name = "devoluciones"
version = "0.1.0"
description = "Sistema web MVP para gestión de devoluciones online"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Equipo Devoluciones" }
]

# Dependencias de runtime (MVP)
dependencies = [
    "Django>=5.0,<6.0",
    "python-decouple>=3.8",
    "djangorestframework>=3.15",
    "django-filter>=24.1",
    "django-cors-headers>=4.3",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-allauth (>=65.13.1,<66.0.0)",
]


[project.optional-dependencies]
# Activar cuando el proyecto escale
async = [
    "celery>=5.4",
    "redis>=5.0",
]

dev = [
    "black>=24.1",
    "isort>=5.13",
    "flake8>=7.0",
    "pytest>=8.0",
    "pytest-django>=4.8",
]


# -----------------------------------------------------------------------------
# HERRAMIENTAS
# -----------------------------------------------------------------------------

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["apps", "config"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_tests.py"]
addopts = "-ra -q"
